SWBIT-4273 #comment Fixes refactor conflicts
